Agora as funções do módulo algorithms são separadas por arquivos.
Criado um arquivo especial para conter todos os testes de unidade. Todos os testes podem ser executados com o comando 'python3 algorithms'.
