+++ algorithms/__main__.py
@@ -1,0 +1,5 @@
+import unittest
+from tests import TestBezout, TestMinimizeCoefficients
+
+if __name__ == "__main__":
+    unittest.main()
